requirements.txt
awscli==1.14.24
botocore==1.8.28
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.7
Django==1.11.3
django-cors-headers==2.1.0
django-model-utils==3.0.0
django-mysql==2.1.0
djangorestframework==3.6.4
docutils==0.14
idna==2.6
jmespath==0.9.3
simple-salesforce==0.73.0
mysqlclient==1.3.12
pyasn1==0.4.2
python-dateutil==2.6.1
python-dotenv==0.8.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
six==1.10.0
urllib3==1.22
git+ssh://bitbucket.org/lapositivatd/proton@0.4.2#egg=proton